split the `Commands` cog into `Miscellaneuos` and `Programming`
